Added TeleopNoA
This teleop mode is identical to teleop, except that it initializes and
starts the rescue arm base in a different position.

This mode is because the rescue arm joint has been physically removed
from the robot, so an alternate teleop is needed to initialize the base
servo to a point where the climbers won't be dropped in the case that we
don't use an autonomous.