Merged Teleop op mode and TeleopNoA op mode
Our climber dropper servo had to be either pointing straight up or
continually powered for the climbers to not fall. In addition, there had
to be no sudden movements if the servo was pointing straight up, or the
climbers might fall. These two requirements present a challenge
considering there might be some games when we run autonomous, and some
games when we do not.

The initial solution to this was to simply duplicate our teleop and
tweak the duplicate so that until the op mode started proper, the servo
would fit the sizing cube. This worked, but the downsides are that we
have to maintain two different versions of near-identical code. What a
mess.

The (slightly) more elegant solution is to use a pseudo-global variable
that can be accessed to determine if an autonomous program was run or
not. This enables us to only have to use one teleop mode.

In addition, 0.267 and 0.538 have been replaced with constants
BASE_RESTING and BASE_VERTICAL, respectively.